Move checksum up into dvm-helper.
