+++ dvm-helper/checksum/checksum.go
