+++ dvm-helper/dvm-helper.go
@@ -11,7 +11,7 @@ import "path/filepath"
 import "regexp"
 import "sort"
 import "strings"
-import "github.com/getcarina/dvm/checksum"
+import "github.com/getcarina/dvm/dvm-helper/checksum"
 import "github.com/fatih/color"
 import "github.com/google/go-github/github"
 import "github.com/codegangsta/cli"
